refactor(test): move permission helper under testhelpers
Why: the cross-platform write-denial package was used exclusively by tests but lived in the production internal namespace, outside the repository's import guard.\n\nWhat: move the package intact beneath internal/testhelpers and update its three test consumers.
